+++ database/seedApi.js
@@ -53,10 +53,19 @@ while (listingIndex < 100) {
   listingIndex++;
 }
 
-Reviews.insertMany(reviewStore, (err, data) => {
+//drops reviews if exist and inserts in else block
+Reviews.deleteMany({id: {$gt: 1} }, (err) => {
   if (err) {
-    console.error('Error: ', err)
+    console.error('Error in deleteMany');
   } else {
-    console.log('Success insertMany: ', data);
+    console.log('Successful deleteMany');
+    //insert many reviews to feed database
+    Reviews.insertMany(reviewStore, (err, data) => {
+      if (err) {
+        console.error('Error: ', err)
+      } else {
+        console.log('Success insertMany: ', data);
+      }
+    });
   }
-});
+})
